made synchronous
