+++ templates/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Pong Game</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
 </head>
 <body>
     <div class="score" id="score">0 - 0</div>
@@ -15,7 +15,7 @@
     </div>
 
     <!-- add script used in end of html  -->
-    <script src="handleOpponent.js"></script>
-    <script src="script.js"></script>
+    <script src="{{ url_for('static', filename='js/handleOpponent.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/script.js') }}"></script>
 </body>
 </html>
